[refactor] 해시맵 이용, 상수 값과 이넘 타입을 키/밸류 쌍으로 이용.
